Refactor conversation title update logic into separate utility function and hook
